backend/fargate/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=backend/fargate/requirements.txt backend/fargate/requirements.in
#
annotated-types==0.7.0
    # via pydantic
boto3==1.42.76
    # via -r backend/fargate/requirements.in
botocore==1.42.76
    # via
    #   boto3
    #   s3transfer
certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
cryptography==46.0.6
    # via google-auth
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.30.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.193.0
    # via
    #   -r backend/fargate/requirements.in
    #   google-generativeai
google-auth==2.49.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-generativeai==0.8.6
    # via -r backend/fargate/requirements.in
googleapis-common-protos==1.73.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.78.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via requests
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
proto-plus==1.27.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.6
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via google-generativeai
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.3.2
    # via httplib2
python-dateutil==2.9.0.post0
    # via botocore
requests==2.33.0
    # via
    #   -r backend/fargate/requirements.in
    #   google-api-core
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
tqdm==4.67.3
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via
    #   botocore
    #   requests
